check_setups: add line number information
Try to find line numbers for every error; they are written to the
output file.  This should enable Jenkins to put comments directly in the
Gerrit change as for pylint warnings.

Do not use readSetup which already catches some errors; we have to
emit these ourselves to be able to add line numbers.

Change-Id: Ic549c6623e8a0df05d394882416305deed3b0c2d
Reviewed-on: https://forge.frm2.tum.de/review/8650
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>