Add an option to start nicos demo with an empty cache
The '-C' option clears the cache (removes the persistent database
files). The gives the opportunity to start always with a clean system
which may be helpful for demonstrations and having no side effects.

At the moment the cache directory will be removed completely but it
could be also simply moved to another name to save the database.

Change-Id: I4d0591e6a15406e1b9ab6e6534adf6de278416f8
Reviewed-on: https://forge.frm2.tum.de/review/8691
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Nikolas Arend <n.arend@fz-juelich.de>