core: improve alias handling with a new setup file entry
(Re)setting aliases in the startupcode works, but is very brittle
if multiple loaded setups set the same alias - the load order then
determines which alias is actually set.

Also, when we remove the setup with a current target, the alias
device currently has a hard time figuring out what to do, especially
if the alias setup is a global one that does not have a default target.

This change introduces a new setup file entry "alias_config" that
solves both issues:

* entries define "desired" alias settings, but they aren't automatically
  applied
* entries define a priority for each alias setting, so that when all
  setups are loaded the one with the highest priority can be selected
* when we remove the setup with the current target, the alias config
  provides a list of reasonable fallbacks, ordered by priority

Change-Id: I2345234e04177d23e6de7bcd9d6e15789844d204
Reviewed-on: https://forge.frm2.tum.de/review/9031
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Georg Brandl <g.brandl@fz-juelich.de>