GUI: disable panels if could not be imported
If the classes for the panels could not be loaded because they are not
installed the will be ignored and not displayed. If the panel is
integrated into a tabbed environment the tab will be suppressed. In
case of a window it will not displayed if the toolbar button is
pressed. A hiding of the toolbutton is at the moment not possible.

Change-Id: Id74ce1675c4f5db14f057947a12835a8ed0ae52f
Fixes: #2325
Reviewed-on: https://forge.frm2.tum.de/review/9204
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>