GUI Status panel: Add an option to abort script execution
For some reason it could be helpful to abort the execution of the
current script (like at TOFTOF) and not waiting until the current line
has been finished. The option 'abortscripts' will change a little bit
the behavior of the stop buttons. It send first a 'pause counting' and
immediately a 'stop' to abort the script execution.

Change-Id: I3dee2e27971a5c0449e293f38325c37159f97e05
Fixes: #2439
Reviewed-on: https://forge.frm2.tum.de/review/9466
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>