NECTAR: add new commands
There is a new tomography command which uses the 'Golden Ratio' to
divide the 360 deg to create information more smoothly. So the user
could stop the data acquisition early if the images give a good chance
to reconstruct.

For adjusting the sample and the camera an additional command is
introduced: alignsample, to put the sample on the origin, or close to it,
to take dark and openbeam images and then move it to the target
position for the tomography itself.

Change-Id: I59f0e247234459858880f27e53a57519e0bad449
Reviewed-on: https://forge.frm2.tum.de/review/9529
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>