DEMO: More realistic setup for the sample changer
In case of loaded 'sans' setup the GUI configuration is more realistic
to the sample changer usage. So the sample changer setups are splitted
into one with 11 positions and one with 22 positions. The GUI config
is changed to react on this. The sample changer is removed from the
sans setup and must be loaded separately.

Change-Id: Ia0eaceaf8f9bcb4bc086d807e3dcfb07da6aa931
Reviewed-on: https://forge.frm2.tum.de/review/9964
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>