core: fix longstanding bug with status propagation
When a device doStatus() method raises an error, the status is supposed
to be ERROR.  Since this try-except was *outside* the get_from_cache call,
the ERROR status was not propagated to the cache however.

This leads to the status not being updated in the device list even when
the poller correctly determines the error state.

Also added a general "except Exception" to make sure that errors always
result in an error state.

Change-Id: I3762e830657702eec37f10c6d546f02e442815be
Fixes: #2602
Reviewed-on: https://forge.frm2.tum.de/review/10283
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Tested-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>