core: add a method for resetting timeout on target change
Use case: see the KWS change -- the target for the chopper freq/phase changes
to zero when the motors are set to "off", but it leads to problem if the
freq/phase is also *moved* there.  So only the target is reset, which also
needs a reset of the _timesout parameter.

Change-Id: I49aaa02b57780baa5cc6d00da1400616ddfddc2c
Reviewed-on: https://forge.frm2.tum.de/review/11164
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>