gui: disallow queueing of certain commands from GUI
For example, if you currently do "Reload setups" and a script is running,
NewSetup() goes into the queue, but the GUI says "Setups reloaded."

For these heavy-weight actions, it's better to disallow them while another
script is running.  For that purpose, we can use the "start" daemon command
but we have to handle the error better than is currently possible, which
is the reason for the changes in the base client.

Change-Id: I0960a5357313b8cac32ddcf853e6522eed1452f8
Reviewed-on: https://forge.frm2.tum.de/review/11266
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>