gui: Allow listed indices also for the display values
The DisplayWidget will also accept a list of indices according to the
other monitor widgets. The plot window could now accept values which
have an index with more than one entry.

The information about the index must be given as a list were the first
entry is for the first dimension, the second for the second dimension,
and so on.

Fixes: #2827
Change-Id: I667c25f0f28732228a5693229e6a02c246bce8f2
Reviewed-on: https://forge.frm2.tum.de/review/11518
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>