SANS-1: Create a copy sink source for detector config files
The instrument scientists want to have for each data file a copy of
the currently used configuration and calibration files. Since the
detector software (QMesyDAQ) is only accessable via TACO and the
TACO interface gives only the filename (but not the path to it back)
the absolute path of the configuration files must be configured.

Change-Id: I4e41be7d199520a3705708dad924c202cc6ebde7
Reviewed-on: https://forge.frm2.tum.de/review/11533
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>