V20 WFM chopper stages and control for distance
This adds control for the two WFM chopper stages at V20.
The motors can be used individually, but the usual mode of
operation is to set a distance between the choppers.
The center of the choppers is always at 0, so that one chopper
is moved to "-distance/2" and the other to "distance/2".

Change-Id: I97261e65f11e0e6dbcefc8090566108076e08546
Fixes: #2785
Reviewed-on: https://forge.frm2.tum.de/review/11588
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>