Added "ETA" (estimated finishing time) to script status panel.
ETA calculation can be enabled via the 'autosimulate'
parameter in the daemon configuration.

ETA display can be enabled via the 'eta' option of the ScriptStatusPanel.
Time format can be adjusted via the 'dateformat' option
of the panel. Default is '%Y/%m/%d %H:%M:%S'.

When enabled:
Every time a script is started/edited the script is simulated and
the duration for the blocks is saved in the Request object.
Every time a script progresses to the next block the ETA is
recalculated and the time send out to the clients.
This is done to account for delays in the last block.

While the simulation is running the status panel will display
a message to indicate to the user that his eta is being calculated.
If the simulation fails the user will be presented with a
corresponding message.

Simresults now send the uuid of the simulated script.
Since the prefix was a workaround to distinguish which simulation
sent the message it is no longer needed.

When the simulation is started by the client a uuid is generated,
which is then applied to the simulated 'Request'(script).
This uuid is then returned with the simresult message and the
client can then determine weather the message is relevant or not.

For ETA simulations triggered by the daemon a quiet flag has
been added.

Change-Id: I3a930038f348db26f5b2664df18498a945f10214
Reviewed-on: https://forge.frm2.tum.de/review/12231
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>