History: Add BareHistory Panel
A Bare HistoryPanel only displays one plot, which can not be
interacted with but will keep updating.
A given option determines which keys will be displayed.
To minimize the space taken by details such as title and
tick labels in gr an option can be set to disable these
details when taking less then the given space.

With the plotsettings option the default display
of BareHistoryPanels can be configured.

Change-Id: Iad7135ab53c67b58245b28f3ec9aa8a1b2214313
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/21579
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Stefan Rainow <s.rainow@fz-juelich.de>