core: disallow device creation from parallel threads
The clients do some queries like "session.instrument.instrument" when connecting,
which can lead to race conditions when NICOS is still loading the initial setup.

This patch forbids parallel execution (of exec/eval) to create new devices.
To enable the clients to query the information reliably, the "setup" event
is connected, which is emitted at the end of the initial setup phase.

Change-Id: Id120b12852b354ce6d091f479777b1736ea37fbc