Fix: BIODIFF rscan shows None in scan results when using microstep motor.
Since for every scan interval the underlying motor will be driven
to the starting point as described in `moveDevices` instead of the
originating motor it is necessary to replace the motor with the
underlying one in the dataset. Otherwise datasets will contain
``None`` values because ``resdict`` in `PointDataset._reslist` will
contain just values for the underlying motor although values for
the originating/microstepping motor are requested.

Change-Id: Iccf7d5eb5f8cf14789d1b48fe0eb048a3c1dde68
Reviewed-on: https://forge.frm2.tum.de/review/12519
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>