Monitor: Add setup specific field display
Each element of the status monitor can be displayed setup
dependent.  If the element is used in inside another group,
the top level group overrides the visibility of the element
in that way if the element would be visible but the group not
the group including the element will be not displayed.

Change-Id: I6dd9c6515e478dff8e5a59ad7f1f47f9d6d69526
Reviewed-on: https://forge.frm2.tum.de/review/16995
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>