TOFTOF: Make the 'ramp' parameter for Iseg modules 'volatile'
Since the Iseg high voltage modules will be switch off during the reactor break
the ramp will be set back to the factory defaults. After switching on, the user
will get back the cached value, which could be completely wrong and in case of
the maximum ramp destroy some of the electronics. So the hardware value should
read every time.

Change-Id: I535b3a92487ca081809ad50b505fd221aa5f3f59
Reviewed-on: https://forge.frm2.tum.de/review/12743
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>