console sink: use dynamic column widths
Up to now, device names > 7 chars lead to a shift of the data/units relative to the names.

With this change, the initial column widths are at least as long as the names and units.
If the data itself is even longer, the widths are dynamically adapted to the longest
entry so that at least following rows have the right width.

Fixes: #3018
Change-Id: Id5e8880f63277402e82d2a0f14fad3a89857befa
Reviewed-on: https://forge.frm2.tum.de/review/12787
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>