Live: Added liveOnly option to LivePanel
By setting the option 'liveonly' one can now hide the last files
list and restrict interaction with the livewidget.

Liveonly defaults to None, which is equivalent to OFF.
In order to display a specific live image liveonly has to be set
to the index of the desired image in the (hidden) listwidget.

Due to the index handling of QListItemView an index which is
too high defaults to the first image.

Change-Id: Idbcc67dd98d7069763630bb39882c52216dbb2f0
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24670
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Stefan Rainow <s.rainow@fz-juelich.de>