V20: Add support for UDP-triggered detector
For a few experiments at V20 there is a detector that
comes with its own control PC and software. It has
a small UDP service which is listening for a trigger
message to start acquiring data and sends a message
when it's done. While not providing much information,
this implementation enables use of all the scan commands.

Change-Id: Icf4729c208d314badb6fc4ab4821b44d81b0eb44
Reviewed-on: https://forge.frm2.tum.de/review/13281
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>