V20: Add Agilent function generator via EPICS
To make integration easier, we tried a different approach
for TCP based devices. Instead of integrating them directly,
there is now a small pcaspy based service that manages
the communication with the device and exposes the device
via channel access. One positive side effect of this is that
no device implementation in lib is required, it's only a setup
file.

Change-Id: I62f521e9a11768a098200745480c46a16b7c64a5
Reviewed-on: https://forge.frm2.tum.de/review/13282
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>