CARESS: Introduce a MuxMotor device
The MuxMotor device is a special implementation of the CARESS motor
device. It has an attached device of type MUX. The only reason for
the attached device is to create the MUX device as a low level device
without an explicit creation and ensure its created before using the
first time (it must be initialized to use it).

Change-Id: Ibb3c10351666558082563f722df9e1caf9e78d1c
Reviewed-on: https://forge.frm2.tum.de/review/13545
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>