SESSION: allow sandboxing simulation runs
Works on Linux >= 2.6.24.

This uses an external binary (needs to be setuid root) that will call
the unshare() system call, which gives the process to create a new
mount and network namespace.  That allows remounting the filesystem
read-only in a chroot, and complete isolation of any network ports that
might be used.

Since TCP is no longer an option, communication to the parent process
is switched to Unix domain sockets (still over ZMQ).

Switched on by setting "sandbox_simulation = true" in nicos.conf.

See https://forge.frm2.tum.de/review/#/c/14075/ for the helper binary.

Change-Id: I6ff0376ce74a204524816dc93bb90bb81960339e
Reviewed-on: https://forge.frm2.tum.de/review/14074
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>