STRESSI: Fix caress data in case of string values
In case of string values the 'READ' entry of the CARESS data file was
corrupted, since the data would not be written into the file.

The CARESS file 'READ' entry accepts only numbers so the string values
must be converted into numbers. If possible the values now will be
converted into corresponding numbers otherwise will be written as '0'.

Change-Id: I89d843c2a87ae8ecd1faf866606117e75530aee2
Reviewed-on: https://forge.frm2.tum.de/review/15011
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>