daemon: close slow/hanging clients more aggressively
The 50ms wait per handler per event is actually quite a lot and
can sum up very quickly (every cache update is an event).

Also, connections where the (100MB!) event queue is full are not
likely to get healthier any time soon.

So, wait a bit longer the first time, but close the connection
if that wait time is exceeded.

Change-Id: I70ca7e2e8534ed100b951c6ad7b72a8a4bfd1f54
Reviewed-on: https://forge.frm2.tum.de/review/15119
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>