ESSIIP: Added Julabo F25 setup
Added a setup for the Julabo F25. The Julabo has a few
features beyond EpicsWindowTimeoutDevice, so a new
device class has been added. Since we have another device
which can be switched on and off remotely, we decided
to give it a go and extract the switching behavior
into a separate class that can be used with EPICS devices.

Change-Id: I8feb2b781cd08a963f68c02f92b660a5c4105e1e
Reviewed-on: https://forge.frm2.tum.de/review/15512
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>