VIRTUAL: motor start/stop/wait fixed
The VirtualMotor device is extensively used in the test suite. Often the
motor speed will be changed and the motor moved without checking the state
for movement. If the speed was switch from != 0 to 0 and the motor was
moving the stop sequence did not work properly, since a possibly running
thread wasn't stopped.

Change-Id: I50b1165ce679dfe6fc3b029f402f49a9067c1f28
Reviewed-on: https://forge.frm2.tum.de/review/15510
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>