EPICS: Improve default doStatus
These changes improve the default doStatus implementation
of EpicsDevice. The EPICS alarm-state and the alarm-severity
are collected for each PV and the most "severe" value is returned
along with the affected PVs.

For the mechanism to work reliably, the control parameters have
to be updated from pyepics each time such a request is made,
because the alarm state and severity are a part of these control
parameters.

Change-Id: Id1645ab9e0a6893d0ad2b7eed3e12aae5b27003c
Reviewed-on: https://forge.frm2.tum.de/review/15533
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Nikhil Biyani <nikhil.biyani@psi.ch>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>