MANTID: Instrument View integration, small sample setup
Mantid must be installed for this to work.

Must be started with at least the following environment variables:
PYTHONPATH=/path/to/mantid/bin QT_API=pyqt INSTRUMENT=nicos_demo.mantid nicos-demo -MEW

May additionally require:
LD_LIBRARY_PATH=/path/to/mantid/lib/paraview-5.3:$LD_LIBRARY_PATH

On a typical install, /path/to/mantid is /opt/Mantid.

Once the InstrumentView is displayed, you'll need to zoom out and
move the view (mouse wheel, [right-]click and drag) a bit to see
the whole instrument. This is just a limitation of this IDF.

Three setups are included:
- pixel: Pixel detector version of SANS2D instrument
- tubes: Tube detector version of SANS2D instrument
- not_viewable: Same as pixel, except the instrument is not viewable

The setups use virtual motors and are purely for testing.

Change-Id: I20470ca92e8ea85f056cef9505df257777c21db8
Reviewed-on: https://forge.frm2.tum.de/review/15682
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Michael Hart <michael.hart@stfc.ac.uk>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>