multiWait: ensure to raise the highest-level error
Ensure to raise an error corresponding to the highest level instead
of the first one encountered.

Some test have been adopted, as we now log all excetions during
multi-wait  (This will give double log output for the final exception
if that is not suppressed later on, but otherwise it is almost
impossible to get the logging timestamps and sequence right).

Change-Id: I5efbd853a308e141f51b2eb81475c2d1c905a3d7
Fixes: #2300
Reviewed-on: https://forge.frm2.tum.de/review/15944
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>