Move test_custom to nicos_mlz and strip prefix
The test_ prefix on test subdirs is not needed for pytest,
so remove it. (but remember: On the test modules the prefix is needed.)

Test_import is  moved to test_simple.

test subdirs for other facilities can follow in later changes.
The same is true for stripping test_ in other subdirs.

Change-Id: Ida604bd3bfd7a3b50a633c97ce12d55811ed0363
Reviewed-on: https://forge.frm2.tum.de/review/15957
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>