CORE: fix devices added to _sdevs but not _adevs, leading to deadlock
When a device had a list of attached devices, and one of the attached
devices failed, the superdev failed to create but was nevertheless added
to the others' _adevs, and the shut down failed with a deadlock.

Now we only add to _sdevs when the whole devlist was created
successfully.

Change-Id: I925e3a4b0585123db2d6d888a799364524a3f3ee
Reviewed-on: https://forge.frm2.tum.de/review/16146
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>