IPC cards: use hardware limits for the reference drive
The IPC Göttingen multichannel cards do not have a reference switch
input. Therefore one of the limit switches must be used. Since there
is a limited range of positions to move the reference drive must take
the whole range but not the absolute limits (config parameter) are
important. The hardware limits (set in some card registers) must be
used.

Change-Id: I55da26a1a892791e99ae8591538807250ca22e12
Reviewed-on: https://forge.frm2.tum.de/review/16376
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>