Fix kscan does not honor live preset.
Previously the detector has been stopped depending on the preset condition and
the `acquireCompleted` callback - whatever matched first. This leaded to
unintentional stopping of the detector. In order to circumvent stopping the
detector in respect to previous presets honor the live preset which results
in counting forever from the point of view of the detector if it would not be
stopped using the `acquireCompleted` callback.

The `Scan` base class pops the live preset in `acquire` because
this is a special preset for unlimited counting which is just
necessary for scans which depends on software synchronisation.
The kscan synchronizes on the movement of the motor.

Change-Id: Ibed383df14a2a1d970ec43d53bbcaf5e849a2e09
Reviewed-on: https://forge.frm2.tum.de/review/16585
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>