REFSANS: Create a single slit device
The single slit device is typically used as the so called
'Zwischenblende'.  It has a differente types of fixed slits with
different heights, shapes, ... . They selection of slits is done
by the mode parameters which initializes a move of the attached
motor so the selected slit is at the same position as the previous
selected one. The position of the slit is the middle of the slit
height.

Change-Id: I05bfae610bfcecfeece7e5684fb8d8bc44008477
Reviewed-on: https://forge.frm2.tum.de/review/16907
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>