MariaDetector + JuelichCounter: Fix live counting.
Due to commit 1c5ff74ca6abfee0184b7290e60ee119673d9cba the timer
channel has not been started anymore if the channel is not in master
mode. Furthermore in master mode the preset 0 be overwritten by 1
to circumvent loss of beam time by accidentially count(t=0).
The TimerChannel has now a ``islive`` parameter which will be set
by the ``MariaDetector``. Furthermore a list of live channels has
been added with this patch named ``lives`` which will be treated as
master channels by the detector.

Change-Id: I51914d5e8c3be87828d781c4b1f8c824bac5c4d3
Reviewed-on: https://forge.frm2.tum.de/review/17112
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>