MARIA ugly module.
In order to conduct some tests during this cycle the ugly module provides
a tango StringIO class which can be used in dry run mode as requested by
the instrument scientists. Devices should be implemented using our
standardized interfaces in tango after this cycle. This module is considered
temporary though.

Change-Id: I0adf39e5ab64e8098c62110626e2d71bf1f94c65
Reviewed-on: https://forge.frm2.tum.de/review/17246
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>