DEMO: Simplify the virtual instrument setup for testing
The setup for testing a virtual instrument, located in
package nicos_demo/v*, with the 'nicos-demo' script can be
configured in a 'nicos.conf', which is not included in git
and located in the checked out source directory.

The nicos.conf file contains these entries:

[nicos]
setup_package = nicos_demo
instrument = vstressi

It avoids to set the whole environment for testing via environment
variables or typing as parameters.
The use of using the well known environment variables is possible
as well.

Change-Id: I1791e293d02e85cfe8e7f3741798605039f81223
Reviewed-on: https://forge.frm2.tum.de/review/17497
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>