SINQ: Implementation of Histogram Memory
Histogram channels are implemented for SINQ/AMOR.
When started or stopped the histogramming is started
or stopped. The data collected is streamed to Kafka
in regular intervals which can be configured using
the saveinterval parameter in detector class. Once
the DAQ is finished the data is also streamed to
Kafka. Live datasinks that shows histogram data
have also been integrated.

Change-Id: Ie191dd1f36acfca6608ac751bb3c48f87e8dd5b1
Reviewed-on: https://forge.frm2.tum.de/review/17525
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Nikhil Biyani <nikhil.biyani@psi.ch>