PUMA: Change the reference driving of multidetector
After reaching the 'compacted' state (all detectors and guides are
touching each other) the spreading of the detectors and the reference
drive of guides has changed. The reference drive of the guides starts
when the detector has moved 'refgap' - 0.2 degrees to avoid that the
reference switch of the guide his hit by touching the neighboring guide.

Change-Id: Ie3f51fb9252b61ff13778cca8a7bc47f18c10e9a
Reviewed-on: https://forge.frm2.tum.de/review/18281
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>