Add private flag to notifiers
The current implementation was inconsistent in handling
notfiers (the Set .. command set only emails on the first
defined Mailer, the gui did fetch from all mailers and set
these all on the first), the watchdog on the otherhand set all
notifiers to the same value.

This change introduces a private flag to mark notifiers as
non-usersettable. As most instruments just define one mailer, the
default of Mailers being public is the less-intrusive change.

Change-Id: I5fb3b56021162408f33b7966791ea85a2ffd9136
Reviewed-on: https://forge.frm2.tum.de/review/18318
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>