PUMA: Add option for the park command of the multi detector
The added parameter 'blocking' (default is True) gives the opportunity
to run the 'park' command not blocking. This is helpful if the PSD
should be used instead of the single counter tubes during a measurement.
Normally the park command is used to move the detectors into a parking
position for dismounting the device.

Change-Id: Ibb90be2b1ab7496659bf6cd8d764b0f4dcb91ce6
Reviewed-on: https://forge.frm2.tum.de/review/18457
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>