device: Forbid assignments of device methods
Sometimes commands like

dev.move = 10

are typed but should not do be allowed, if the member is a method of
the device.

We also introduce a warning if any other attribute is set that is
not a property. This may break on wild instrument code so is not
made an error immediately.

Change-Id: I8793e050ec6014a353c26812890af7fac21ff3b4
Reviewed-on: https://forge.frm2.tum.de/review/18833
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>