livewidget: remove empty string datatype.
This is a relict of the Qwt implementation which has been used
as a starting point for the GR version. Nevertheless the empty type
is not necessary anymore as the ``DATATYPES`` constant is just checked
against the normalized types (line 409) which are retrieved using
``numpy.dtype(dtype).str`` which will raise an `TypeError` on ``numpy.dtype('')``.
The empty ``dtype`` check necessary for `notifyDataFile` is covered in line 405
by a conditional on the non-normalized ``dtype`` variable.

Change-Id: Iac90e7f9d931f499731f2df14f71f050f7f852ab
Reviewed-on: https://forge.frm2.tum.de/review/19181
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>