SPHERES: add Sequencer Mixin to the doppler
Moved doppler status into Doppler device.
Speeds are now compared in the doppler device as
opposed to the acq doppler.
switch off is properly recognized as speed 0.

When the doppler changes speeds the detector needs some time
to adapt to this speed. In order to keep the doppler busy
during this transition the speed changing process has been
moved to a sequence which at the end waits until the detector
is finished adjusting.

Change-Id: I890ff17b722b978673bd6639055d1c5c9cdcd46d
Reviewed-on: https://forge.frm2.tum.de/review/19947
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Stefan Rainow <s.rainow@fz-juelich.de>