Improve parseDuration and move it to nicos.utils
Allow non-integer values for formatted strings, number-like strings
and timedelta objects. Move it to nicos utils and raise standard
exceptions.

Change in behaviour: now raises a ValueError for non-parseable
input (this is an error in user input typically) or a TypeError for
wrong input types(this is a programming error).
These should be caught on usage.

Change-Id: I729ab04407a9bd987b5363189311193802a75f08
Reviewed-on: https://forge.frm2.tum.de/review/19591
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Stefan Rainow <s.rainow@fz-juelich.de>