DataSet: Add finish_property
Finish properties are properties which use the provided read
function to get the requested values until their owner's finished
flag is set. Then they become 'lazy'.

Properties like 'detvaluelist' in PointDatasets need to be
updated in case of multiple saves per file like during an
intermediate save and the subsequent final save.
They are calculated from results which will be
stripped at the end of the scan technically resetting them
in the process.
To avoid this unwanted behavior the properties in question
are made lazy at the first call after the finished flag has
been set.
These calls are explicitly made before striping the results.

The porperties in question of the PointDataset have been adjusted.

Change-Id: Ia4571cd4565c1fbece57b31825a6b713205a2d3e
Reviewed-on: https://forge.frm2.tum.de/review/20313
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>