REFSANS: Reworking of safety system
Complete rewriting of the 'Group' devices. The configuration given via
the 'bit lis't names (defined in code) and the 'ok mask' which gives the
expected bit mask.

Read value is either True or False, the status gives a WARN state if the
bit map is not correctly set and tells not correctly set bit names.

Change-Id: I4668148f8ab3b6393a651cfb42a8335aac66dc90
Reviewed-on: https://forge.frm2.tum.de/review/20479
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>