REFSANS: Fix not initializing chopper disc2 device
Due to a limitation of the hardware the position of the chopper
disc2 position can be 0 but it is not a valid position. It indicates
that the hardware is not referenced. The parameter read raises always
an error, so the device can't be initialised.

To overcome the problem the parameter type for the hardware is extended
to the 0 position, but the writing of 0 raises an error.

Change-Id: I35b3c3702c7ad608a807088c8c16f1835f2bb9f7
Reviewed-on: https://forge.frm2.tum.de/review/20726
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>