REFSANS: Remove 'nok_length' parameter from optic devices
The 'nok_length' parameter is remove due to the doubled information:

length = nok_end - nok_start

A volatile parameter is introduced to have an easy access to the
length.

For the double slit the nok_start and nok_end parameters are taken
from the nok_start of the attached reactor side slit and the nok_end
from the sample side slit.

Change-Id: I3d001057216964d79e6b8c56447feab09b38ac74
Reviewed-on: https://forge.frm2.tum.de/review/20910
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>