Core: Fixed issue with ns10 flatbuffer creation
Verifying the NICOS created ns10 flatbuffer using the C++ library
fails. It appears from testing that the flatbuffer fields need to be
added in the reverse order of the schema to pass verification.

This is probably a bug in the Python flatbuffers implementation (and,
perhaps, specific to Python3). When constructing it in C++ the order
does not affect verification.

This patch constructs the flatbuffer in the reverse order.

Change-Id: I6f7518f2e2f0aea4eb787fb5f6085acf3d6f6932