GUI: allow different interface in derived class
In the default implementation the ui file for the panel is hard-coded in
the __init__::loadUi. This makes not possible to have a derived panel that
uses a different ui file.
This patch makes more easy to have a different ui in the children (but
doesn't give the full flexibility that one can have if the parameter
could be se in the guiconfig.py).

Change-Id: I1f4776f11a05478e89d9d151000670c2977c6fe6
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/21719
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se >