monitor: allow defining monitor blocks in device setups
This tries to get rid of the rampant code duplication in
different monitor setups (gui, html, office-raspi, ...)

Even if there is still some difference in blocks, they are all
defined in the same setup (the one providing the devices too)
and can use common parts in variables.

It is also much easier to remember adapting the monitor config
when changing device setups.

Change-Id: Ic069094529f2b7b3dd6ed79435ce55cbeacae9bd
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/21906
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>