ESS: split EpicsKafkaForwarder in a monitor and a controller
We want to be able to use the EpicsKafkaForwarder just as a monitor for
the status of the forwarder. In this case it should not issue commands
to the ESS forward-epics-to-kafka, but jst make sure that is is
forwardind and how many PVs are forwarded.
If one wants the NICOS EpicsKafkaForwarder to control (e.g. adding new
PVs) the ESS forwarder, the new EpicsKafkaForwarderControl can be used
(as a device attached to the EpicsKafkaForwarder)

Change-Id: I7aef12df5ab72ad55b4a7713f4441c43ece6bec5
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22210
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se >